+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.core/src/main/java/org/eclipse/lsp4jakarta/jdt/internal/interceptor/InterceptorDiagnosticsParticipant.java
@@ -36,7 +36,7 @@ import org.eclipse.lsp4jakarta.jdt.core.utils.PositionUtils;
 import org.eclipse.lsp4jakarta.jdt.internal.DiagnosticUtils;
 import org.eclipse.lsp4jakarta.jdt.internal.Messages;
 import org.eclipse.lsp4jakarta.jdt.internal.core.ls.JDTUtilsLSImpl;
-import org.eclipse.lsp4jakarta.jdt.core.java.diagnostics.helper.ConstructorInfoDiagnosticHelper;
+import org.eclipse.lsp4jakarta.jdt.core.java.diagnostics.helpers.ConstructorInfoDiagnosticHelper;
 
 /**
  * Interceptor diagnostic participant that manages the use of @Interceptor annotation.
